backend/requirements.txt
pyttsx3
SpeechRecognition
PyAudio
pvporcupine
python-dotenv
huggingface-hub
python-osc
fastapi>=0.68.0
uvicorn>=0.15.0
whisper-openai
transformers
torch
# Optional: if you want to use CUDA with torch
# torch --index-url https://download.pytorch.org/whl/cu118